adds a yt video player
adds $search to search youtube, download a video and play it. This is very buggy. you must be in a music channel to use. once a video starts playing there is no way to stop it. once the video ends the bot crashes and must be restarted. once it is used the music file must be manually deleated.