bugfix
This attempts to fix a bug with audio where the queue system will break and no more playback will be possible until a restart. i have reproduced this through normal use, however, I  was unable to reproduce this during testing. I added a few lines here and there to fix a couple of reasons why i think this may happen.